moved some controller interfaces into the correct places
