--- XRTK-Core/Packages/com.xrtk.core/Interfaces/InputSystem/Controllers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a4d0fd8d66259047ad195a022d7f3e5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
