+++ XRTK-Core/Packages/com.xrtk.core/Interfaces/Providers/Controllers/ISimulatedControllerDataProvider.cs
@@ -2,9 +2,8 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEngine;
-using XRTK.Interfaces.Providers.Controllers;
 
-namespace XRTK.Interfaces.InputSystem.Controllers.Hands
+namespace XRTK.Interfaces.Providers.Controllers
 {
     public interface ISimulatedControllerDataProvider : IMixedRealityControllerDataProvider
     {
