+++ XRTK-Core/Packages/com.xrtk.core/Providers/Controllers/BaseController.cs
@@ -10,6 +10,7 @@ using XRTK.Definitions.InputSystem;
 using XRTK.Definitions.Utilities;
 using XRTK.Extensions;
 using XRTK.Interfaces.InputSystem;
+using XRTK.Interfaces.InputSystem.Handlers;
 using XRTK.Interfaces.Providers.Controllers;
 using XRTK.Services;
 using XRTK.Utilities.Gltf.Schema;
