+++ XRTK-Core/Packages/com.xrtk.core/Providers/Controllers/Hands/BaseHandController.cs
@@ -7,8 +7,8 @@ using XRTK.Definitions.Controllers.Hands;
 using XRTK.Definitions.Devices;
 using XRTK.Definitions.Utilities;
 using XRTK.Interfaces.InputSystem;
-using XRTK.Interfaces.InputSystem.Controllers.Hands;
 using XRTK.Interfaces.Providers.Controllers;
+using XRTK.Interfaces.Providers.Controllers.Hands;
 using XRTK.Services;
 
 namespace XRTK.Providers.Controllers.Hands
