+++ XRTK-Core/Packages/com.xrtk.core/Providers/Controllers/Hands/BaseHandControllerDataProvider.cs
@@ -2,8 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using XRTK.Definitions.Controllers.Hands;
-using XRTK.Interfaces.InputSystem.Controllers.Hands;
-using XRTK.Interfaces.Providers.Controllers;
+using XRTK.Interfaces.Providers.Controllers.Hands;
 using XRTK.Services;
 
 namespace XRTK.Providers.Controllers.Hands
