+++ XRTK-Core/Packages/com.xrtk.core/Providers/Controllers/Simulation/Hands/SimulatedHandController.cs
@@ -9,8 +9,8 @@ using XRTK.Definitions.InputSystem;
 using XRTK.Definitions.Utilities;
 using XRTK.Extensions;
 using XRTK.Interfaces.InputSystem;
-using XRTK.Interfaces.InputSystem.Controllers.Hands;
 using XRTK.Interfaces.Providers.Controllers;
+using XRTK.Interfaces.Providers.Controllers.Hands;
 using XRTK.Providers.Controllers.Hands;
 using XRTK.Services;
 using XRTK.Utilities;
